refactoring to remove unnecessary code
