+++ eCRF_frontend/src/components/StudyDataEntryComponent.vue
